requirements.txt
git+ssh://git@github.com/rapptz/discord.py@async
pyyaml
youtube-dl
